Add IMU polling, state monitoring, add CameraWorkerInfo dataclass, move process to within IMUManager

+++ src/modules/imu/imu_manager.py
@@ -16,6 +16,9 @@ class IMUManager:
         self.stop_event = stop_event
         self.imu_data = imu_data
         self.__logger = logging.getLogger(__name__)
+        self.imu_process = None
+        self.imu_worker = None
+
     def start_imu_worker(self, imu_data):
         """
         Initializes IMU worker using shared memory
@@ -29,7 +32,24 @@ class IMUManager:
         Stops all activate camera processes and terminates gracefully
         """
         self.__logger.info(f"Stopping IMU process {self.imu_process}")
-        # Tells each worker to exit the stream_data loop
-        self.stop_event.set()
-        self.imu_worker.stop_imu()
-        self.imu_process.terminate()
+                
+        if self.imu_process:
+            # Tells each worker to exit the stream_data loop
+            self.stop_event.set()
+            self.imu_worker.stop_imu()
+
+            self.imu_process.join(timeout=1.0)
+
+            if self.imu_process.is_alive():
+                self.__logger.warning("IMU process still alive. Terminating...")
+                self.imu_process.terminate()
+
+            self.__logger.info("IMU worker process stopped")
+
+    def is_running(self):
+        """
+        Returns true if IMU process is still alive
+        """
+        if self.imu_process:
+            return self.imu_process.is_alive()
+        return False
